+++ src/store/feedPost/Selectors.ts
@@ -1,0 +1,3 @@
+import { IState } from 'interfaces/stateInterface';
+
+export const posts = (store: IState) => store.feedPost.posts;
