--- src/store/post/Selectors.ts
@@ -1,3 +1,0 @@
-import { IState } from 'interfaces/stateInterface';
-
-export const posts = (store: IState) => store.post.posts;
